requirements/development.txt
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
coverage==5.4
cryptography==3.4.5
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==3.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pymongo==3.11.3
pyparsing==2.4.7
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
Flask-Principal==0.4.0
Flask-Security-Too==4.0.0

